Set up Gradle support and Checkstyle plugin
