+++ src/main/data/duke.txt
@@ -1,3 +1,3 @@
 T | 0 | read book
-D | 0 | return book  | 2020-12-08
+D | 0 | return book | 2020-12-08
 T | 0 | throw trash
