+++ src/main/java/duke/command/DoneCommand.java
@@ -1,4 +1,4 @@
-package duke.Command;
+package duke.command;
 
 import duke.DukeException;
 import duke.Storage;
