+++ src/main/java/duke/command/ListCommand.java
@@ -1,4 +1,4 @@
-package duke.Command;
+package duke.command;
 
 import duke.DukeException;
 import duke.Storage;
@@ -6,13 +6,12 @@ import duke.TaskList;
 import duke.Ui;
 
 /**
- * Represents the user command to exit the program.
+ * Represents the user command to output all the tasks in the TaskList.
  */
-public class ExitCommand extends Command {
+public class ListCommand extends Command {
 
     /**
-     * Executes the exit command. A goodbye message is shown to the user, then the tasks
-     * from the TaskList are written and saved into the file.
+     * Executes the list command. Each task in the TaskList is printed in order.
      *
      * @param taskList A TaskList object that contains an arraylist of Task objects.
      * @param ui A Ui object that deals with interactions with the user.
@@ -20,16 +19,15 @@ public class ExitCommand extends Command {
      */
     @Override
     public void runCommand(TaskList taskList, Ui ui, Storage storage) throws DukeException {
-        ui.showGoodbye();
-        storage.save(taskList);
+        ui.printTasks(taskList);
     }
 
     /**
      * Indicates if the command ends the program after executing.
-     * @return true
+     * @return false
      */
     @Override
     public boolean isExit() {
-        return true;
+        return false;
     }
 }
