+++ src/test/java/duke/DukeTest.java
@@ -1,13 +1,15 @@
 package duke;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+
 
 public class DukeTest {
     @Test
     public void dummyTest() {
-        assertEquals(2,2);
+        assertEquals(2, 2);
     }
