+++ src/test/java/duke/ParserTest.java
@@ -1,10 +1,4 @@
 package duke;
 
-import duke.Command.ListCommand;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertSame;
-
 public class ParserTest {
 }
