+++ src/test/java/duke/UiTest.java
@@ -1,8 +1,4 @@
 package duke;
 
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
 public class UiTest {
 }
