+++ src/test/java/duke/command/AddCommandTest.java
@@ -1,4 +1,4 @@
-package duke.Command;
+package duke.command;
 
 public class AddCommandTest {
 }
